refact: 좋아요 버튼 `LikeButton` 컴포넌트로 분리
- DetailView 컴포넌트에서 좋아요 버튼을 LikeButton 컴포넌트로 분리
- #8
